Class WP_Importer moved to Devtronic\FreshPress\Components\Misc\Importer
